Share thread Pause code via composition instead
