+++ src/Spice86.Core/Emulator/Devices/Sound/OPL3FM.cs
@@ -4,6 +4,7 @@ using Spice86.Core.Backend.Audio;
 using Spice86.Core.Emulator.CPU;
 using Spice86.Core.Emulator.Devices.Sound.Ymf262Emu;
 using Spice86.Core.Emulator.IOPorts;
+using Spice86.Core.Emulator.Pause;
 using Spice86.Core.Emulator.Sound;
 using Spice86.Core.Emulator.Devices.Sound.Ymf262Emu;
 using Spice86.Shared.Interfaces;
@@ -13,7 +14,7 @@ using System;
 /// <summary>
 /// Virtual device which emulates OPL3 FM sound.
 /// </summary>
-public sealed class OPL3FM : PauseableDevice, IDisposable {
+public sealed class OPL3FM : DefaultIOPortHandler, IPauseable, IDisposable {
     private const byte Timer1Mask = 0xC0;
     private const byte Timer2Mask = 0xA0;
 
@@ -29,6 +30,10 @@ public sealed class OPL3FM : PauseableDevice, IDisposable {
     private byte _timerControlByte;
 
     private bool _disposed;
+    private bool _isPaused;
+
+    /// <inheritdoc/>
+    public bool IsPaused { get => _isPaused; set => _isPaused = value; }
 
     /// <summary>
     /// Initializes a new instance of the OPL3 FM synth chip.
@@ -141,7 +146,7 @@ public sealed class OPL3FM : PauseableDevice, IDisposable {
         Span<float> playBuffer = stackalloc float[length];
         FillBuffer(buffer, playBuffer, expandToStereo);
         while (!_endThread) {
-            SleepWhilePaused();
+            ThreadPause.SleepWhilePaused(ref _isPaused);
             _audioPlayer.WriteFullBuffer(playBuffer);
             FillBuffer(buffer, playBuffer, expandToStereo);
         }
