+++ src/Spice86.Core/Emulator/Devices/Sound/SoundBlaster.cs
@@ -8,6 +8,7 @@ using Spice86.Core.Emulator.Devices.DirectMemoryAccess;
 using Spice86.Core.Emulator.Devices.ExternalInput;
 using Spice86.Core.Emulator.IOPorts;
 using Spice86.Core.Emulator.Memory;
+using Spice86.Core.Emulator.Pause;
 using Spice86.Core.Emulator.Sound;
 using Spice86.Core.Emulator.Sound.Blaster;
 using Spice86.Shared.Interfaces;
@@ -19,7 +20,7 @@ using System.Threading;
 /// Sound blaster implementation. <br/>
 /// http://www.fysnet.net/detectsb.htm
 /// </summary>
-public sealed class SoundBlaster : PauseableDevice, IDmaDevice8, IDmaDevice16, IRequestInterrupt, IBlasterEnvVarProvider, IDisposable {
+public sealed class SoundBlaster : DefaultIOPortHandler, IPauseable, IDmaDevice8, IDmaDevice16, IRequestInterrupt, IBlasterEnvVarProvider, IDisposable {
     /// <summary>
     /// The port number for checking if data is available to be read from the DSP.
     /// </summary>
@@ -137,6 +138,7 @@ public sealed class SoundBlaster : PauseableDevice, IDmaDevice8, IDmaDevice16, I
     private int _pauseDuration;
     private BlasterState _blasterState;
     private bool _playbackStarted;
+    private bool _isPaused;
     private readonly DualPic _dualPic;
     private readonly IGui? _gui;
     private readonly DmaController _dmaController;
@@ -291,6 +293,9 @@ public sealed class SoundBlaster : PauseableDevice, IDmaDevice8, IDmaDevice16, I
     /// </summary>
     public FrozenSet<int> OutputPorts => new int[] { DspPorts.DspReset, DspPorts.DspWrite, DspPorts.MixerAddress }.ToFrozenSet();
 
+    /// <inheritdoc/>
+    public bool IsPaused { get => _isPaused; set => _isPaused = value; }
+
     /// <inheritdoc />
     public void Dispose() {
         // Do not change this code. Put cleanup code in 'Dispose(bool disposing)' method
@@ -352,7 +357,7 @@ public sealed class SoundBlaster : PauseableDevice, IDmaDevice8, IDmaDevice16, I
         int sampleRate = player.Format.SampleRate;
 
         while (!_endPlayback) {
-            SleepWhilePaused();
+            ThreadPause.SleepWhilePaused(ref _isPaused);
             _dsp.Read(buffer);
             int length = Resample(buffer, sampleRate, writeBuffer);
             player.WriteFullBuffer(writeBuffer.AsSpan(0, length));
