+++ src/Spice86.Core/Emulator/Sound/Midi/MidiDevice.cs
@@ -7,13 +7,17 @@ using System;
 /// <summary>
 /// The base class for all classes talking to an external MIDI device.
 /// </summary>
-internal abstract class MidiDevice : Pauseable, IDisposable {
+internal abstract class MidiDevice : IPauseable, IDisposable {
     private uint _currentMessage;
     private uint _bytesReceived;
     private uint _bytesExpected;
     private byte[] _currentSysex = new byte[128];
     private int _sysexIndex = -1;
+    protected bool _isPaused;
     private static readonly uint[] MessageLength = { 3, 3, 3, 3, 2, 2, 3, 1 };
+
+    /// <inheritdoc/>
+    public bool IsPaused { get => _isPaused; set => _isPaused = value; }
 
     /// <summary>
     /// Sends a byte to the MIDI device.
